fix: 重新调整顶栏结构，按组件拆分逻辑，通过store管理组件间状态  #8 #11
