+++ package.json
@@ -2,9 +2,9 @@
   "name": "bewly-cat",
   "displayName": "BewlyCat",
   "type": "module",
-  "version": "0.0.7",
+  "version": "0.0.8",
   "private": true,
-  "packageManager": "pnpm@9.5.0",
+  "packageManager": "pnpm@10.8.0",
   "description": "Just make a few small changes to your Bilibili homepage.",
   "homepage": "https://github.com/keleus/BewlyCat",
   "scripts": {
