+++ src/components/TopBar/types.ts
@@ -121,3 +121,12 @@ export interface TopBarItemElements {
 export interface TopBarTransformers {
   [key: string]: Ref<any>
 }
+
+export interface TopBarInteraction {
+  popupVisible: PopupVisibleState
+  currentClickedTopBarItem: Ref<string | null>
+  handleClickTopBarItem: (event: MouseEvent, key: string) => void
+  setupTopBarItems: () => any
+  setupClickOutside: () => void
+  registerTopBarItem: (key: string, element: HTMLElement | null, transformer: any) => void
+}
