+++ src/contentScripts/index.ts
@@ -1,6 +1,7 @@
 import '~/styles'
 import 'uno.css'
 
+import { createPinia } from 'pinia'
 import { createApp } from 'vue'
 
 import { useDark } from '~/composables/useDark'
@@ -398,6 +399,7 @@ function injectApp() {
   document.body.appendChild(container)
 
   const app = createApp(App)
+  app.use(createPinia())
   setupApp(app)
   app.mount(root)
 }
